+++ index.js
@@ -42,11 +42,16 @@ const{
 	// Upload indexes
 	await uploadResources(indexesDir, "indexes");
 
-	// Upload functions
-	await uploadResources(fnsDir, "functions");
+	// Upload functions without their role property.
+	// This solves a problem where the fauna would throw an invalid reference error
+	// when referencing user defined roles before they exist.
+	await uploadResources(fnsDir, "functions", { fnsWithRoles: false });
 
 	// Upload roles
 	await uploadResources(rolesDir, "roles");
+
+	// Upload functions with roles
+	await uploadResources(fnsDir, "functions", { fnsWithRoles: true })
 
 	// Upload data
 	await uploadResources(dataDir, "data");